requirements.txt
diffusers==0.31.0
torch
pillow
tqdm
